# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Local business discovery engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/review-queue:
    get:
      operationId: getReviewQueue
      tags: [businesses]
      summary: List businesses that should be manually reviewed next
      parameters:
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: targetMarket
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Prioritized review queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewQueueItem"
...
  /businesses/{id}/outreach:
    get:
      operationId: getBusinessOutreach
      tags: [businesses]
      summary: Get protected outreach fields for one business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Outreach details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessOutreach"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBusinessOutreach
      tags: [businesses]
      summary: Update protected outreach fields for one business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessOutreachRequest"
      responses:
        "200":
          description: Updated outreach details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessOutreach"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/sources:
    get:
      operationId: getBusinessSources
      tags: [businesses]
      summary: List tracked sources for one business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sources for one business
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessSource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/contact-candidates:
    get:
      operationId: getBusinessContactCandidates
      tags: [businesses]
      summary: List contact candidates derived for one business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact candidates for one business
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactCandidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/outreach-events:
    get:
      operationId: getBusinessOutreachEvents
      tags: [businesses]
      summary: List outreach audit events for one business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Outreach audit event timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutreachEvent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/contact-candidates/{candidateId}:
    patch:
      operationId: updateBusinessContactCandidate
      tags: [businesses]
      summary: Update review state or primary flag for one contact candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactCandidateUpdateRequest"
      responses:
        "200":
          description: Updated contact candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactCandidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /research/views:
    get:
      operationId: getResearchViews
      tags: [research]
      summary: List saved research views
      responses:
        "200":
          description: Saved views
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchView"
    post:
      operationId: createResearchView
      tags: [research]
      summary: Create a saved research view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchViewPayload"
      responses:
        "201":
          description: Saved view created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchView"
        "400":
          description: Invalid saved view request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /research/views/{id}:
    patch:
      operationId: updateResearchView
      tags: [research]
      summary: Update a saved research view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchViewPayload"
      responses:
        "200":
          description: Saved view updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchView"
        "404":
          description: Saved view not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResearchView
      tags: [research]
      summary: Delete a saved research view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Saved view deleted
...
  /outreach/dashboard:
    get:
      operationId: getOutreachDashboard
      tags: [outreach]
      summary: Outreach KPI dashboard metrics
      parameters:
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: targetMarket
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Outreach dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachDashboardResponse"
  /outreach/pipeline:
    get:
      operationId: getOutreachPipeline
      tags: [outreach]
      summary: Follow-up board items and urgency buckets
      parameters:
        - name: horizonDays
          in: query
          schema:
            type: integer
            default: 7
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: targetMarket
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Outreach pipeline board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachPipelineResponse"
...
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        categorySlug:
          type: string
        name:
          type: string
        slug:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        addressLine:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        osmId:
          type: ["string", "null"]
        osmType:
          type: ["string", "null"]
        googlePlaceId:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        rating:
          type: ["string", "null"]
        userRatingsTotal:
          type: ["integer", "null"]
        hasWebsite:
          type: boolean
        hasPhone:
          type: boolean
        enrichmentStatus:
          type: string
        targetMarket:
          type: ["string", "null"]
        engineType:
          type: ["string", "null"]
        targetType:
          type: ["string", "null"]
        targetCluster:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        discoveryStatus:
          type: ["string", "null"]
        qualificationStatus:
          type: ["string", "null"]
        contactabilityStatus:
          type: ["string", "null"]
        rankingStatus:
          type: ["string", "null"]
        sourceHealth:
          type: ["string", "null"]
        contactReadiness:
          type: ["string", "null"]
        relevanceScore:
          type: ["integer", "null"]
        contactabilityScore:
          type: ["integer", "null"]
        confidenceScore:
          type: ["integer", "null"]
        freshnessScore:
          type: ["integer", "null"]
        priorityScore:
          type: ["integer", "null"]
        researchScore:
          type: ["integer", "null"]
        economicValueScore:
          type: ["integer", "null"]
        strategicValueScore:
          type: ["integer", "null"]
        referralValueScore:
          type: ["integer", "null"]
        prestigeValueScore:
          type: ["integer", "null"]
        continuityRevenuePotential:
          type: ["integer", "null"]
        offerFitScore:
          type: ["integer", "null"]
        relationshipPathScore:
          type: ["integer", "null"]
        actionabilityScore:
          type: ["integer", "null"]
        officialSourceCount:
          type: ["integer", "null"]
        successfulSourceCount:
          type: ["integer", "null"]
        failedSourceCount:
          type: ["integer", "null"]
        primarySourceId:
          type: ["integer", "null"]
        primaryContactCandidateId:
          type: ["integer", "null"]
        bestOfferId:
          type: ["integer", "null"]
        bestNarrativeId:
          type: ["integer", "null"]
        secondaryNarrativeId:
          type: ["integer", "null"]
        bestCredibilityAssetId:
          type: ["integer", "null"]
        bestCaseStudyId:
          type: ["integer", "null"]
        proofAngle:
          type: ["string", "null"]
        riskReductionReason:
          type: ["string", "null"]
        toneOfApproach:
          type: ["string", "null"]
        recommendedPitchAngle:
          type: ["string", "null"]
        nextBestContactWindow:
          type: ["string", "null"]
        accountTier:
          type: ["string", "null"]
        seasonalityFit:
          type: ["string", "null"]
        warmPathExists:
          type: boolean
        readyForRelationship:
          type: boolean
        readyForInstitutionalPitch:
          type: boolean
        prestigeWatchlist:
          type: boolean
        cultivationRequired:
          type: boolean
        readyForOutreach:
          type: boolean
        reviewRequired:
          type: boolean
        reviewReason:
          type: ["string", "null"]
        topGap:
          type: ["string", "null"]
        recommendedNextStep:
          type: ["string", "null"]
        lastResearchAt:
          type: ["string", "null"]
        nextResearchAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - categorySlug
        - name
        - slug
        - latitude
        - longitude
        - hasWebsite
        - hasPhone
        - enrichmentStatus
        - warmPathExists
        - readyForRelationship
        - readyForInstitutionalPitch
        - prestigeWatchlist
        - cultivationRequired
        - readyForOutreach
        - reviewRequired
        - createdAt
        - updatedAt
...
    BusinessSource:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        sourceType:
          type: string
        sourceUrl:
          type: string
        sourceDomain:
          type: ["string", "null"]
        discoveredVia:
          type: ["string", "null"]
        fetchStatus:
          type: string
        lastFetchedAt:
          type: ["string", "null"]
        contentHash:
          type: ["string", "null"]
        httpStatus:
          type: ["integer", "null"]
        isOfficial:
          type: boolean
        sourcePriority:
          type: integer
        usefulnessScore:
          type: ["integer", "null"]
        nextFetchAt:
          type: ["string", "null"]
        freshnessStatus:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - sourceType
        - sourceUrl
        - fetchStatus
        - isOfficial
        - sourcePriority
        - freshnessStatus
        - createdAt
        - updatedAt
    ReviewQueueItem:
      type: object
      properties:
        business:
          $ref: "#/components/schemas/Business"
        reasons:
          type: array
          items:
            type: string
        priorityScore:
          type: integer
        sourceCount:
          type: integer
        officialSourceCount:
          type: integer
        failedSourceCount:
          type: integer
      required:
        - business
        - reasons
        - priorityScore
        - sourceCount
        - officialSourceCount
        - failedSourceCount
    BusinessOutreach:
      type: object
      properties:
        businessId:
          type: integer
        outreachStatus:
          type: string
        contactName:
          type: ["string", "null"]
        contactRole:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        lastContactDate:
          type: ["string", "null"]
        nextActionDate:
          type: ["string", "null"]
        assignedArtist:
          type: ["string", "null"]
        assignedArtistSource:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        targetMarket:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        warmConnection:
          type: ["string", "null"]
        readyForOutreach:
          type: boolean
        reviewRequired:
          type: boolean
        reviewReason:
          type: ["string", "null"]
        priorityScore:
          type: ["integer", "null"]
        updatedAt:
          type: string
      required:
        - businessId
        - outreachStatus
        - readyForOutreach
        - reviewRequired
        - updatedAt
    UpdateBusinessOutreachRequest:
      type: object
      properties:
        outreachStatus:
          type: string
        contactName:
          type: ["string", "null"]
        contactRole:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        lastContactDate:
          type: ["string", "null"]
        nextActionDate:
          type: ["string", "null"]
        assignedArtist:
          type: ["string", "null"]
        assignedArtistSource:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        targetMarket:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        warmConnection:
          type: ["string", "null"]
    ContactCandidate:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        fullName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        contactType:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        contactUrl:
          type: ["string", "null"]
        sourceUrl:
          type: string
        sourceType:
          type: string
        confidenceScore:
          type: string
        isPrimary:
          type: boolean
        isPersonalData:
          type: boolean
        lastVerifiedAt:
          type: ["string", "null"]
        verificationStatus:
          type: string
        isReachable:
          type: boolean
        isDecisionMakerLikely:
          type: boolean
        channelPriority:
          type: integer
        sourcePriority:
          type: integer
        nextVerificationAt:
          type: ["string", "null"]
        reviewStatus:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - contactType
        - sourceUrl
        - sourceType
        - confidenceScore
        - isPrimary
        - isPersonalData
        - verificationStatus
        - isReachable
        - isDecisionMakerLikely
        - channelPriority
        - sourcePriority
        - reviewStatus
        - createdAt
        - updatedAt
    OutreachEvent:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        eventType:
          type: string
        entityType:
          type: string
        entityId:
          type: ["integer", "null"]
        actorType:
          type: string
        summary:
          type: string
        changedFields:
          type: array
          items:
            type: string
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required:
        - id
        - businessId
        - eventType
        - entityType
        - actorType
        - summary
        - changedFields
        - payload
        - createdAt
...
    ResearchView:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scope:
          type: string
        isDefault:
          type: boolean
        filtersJson:
          type: object
          additionalProperties: true
        sortJson:
          type: object
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, scope, isDefault, filtersJson, sortJson]
    ResearchViewPayload:
      type: object
      properties:
        name:
          type: string
        scope:
          type: string
        isDefault:
          type: boolean
        filtersJson:
          type: object
          additionalProperties: true
        sortJson:
          type: object
          additionalProperties: true
      required: [name]
    ContactCandidateUpdateRequest:
      type: object
      properties:
        reviewStatus:
          type: string
        isPrimary:
          type: boolean
...
    OutreachPipelineResponse:
      type: object
      properties:
        today:
          type: string
        horizonDays:
          type: integer
        total:
          type: integer
        summary:
          $ref: "#/components/schemas/OutreachPipelineSummary"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OutreachPipelineItem"
      required: [today, horizonDays, total, summary, items]
    OutreachDashboardResponse:
      type: object
      properties:
        totalTargets:
          type: integer
        contacted:
          type: integer
        positiveResponses:
          type: integer
        responseRate:
          type: number
        overdueFollowUps:
          type: integer
        inProgress:
          type: integer
        interested:
          type: integer
        urgentThisWeek:
          type: array
          items:
            $ref: "#/components/schemas/OutreachPipelineItem"
      required:
        - totalTargets
        - contacted
        - positiveResponses
        - responseRate
        - overdueFollowUps
        - inProgress
        - interested
        - urgentThisWeek
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]